rename Dataset to ContinuumDataset for better distinction and I moved it to dataset folder to split dataset classes from loader
